Split out views
Since will be adding more routes splitting this out to make it more manageable
